Remove "select all" and "select none" buttons for filters.
During user testing at the Digital Ready Maker Party, people un-checked
a box and expected that category to show on the map, which wasn't what
happened. Now we start out with nothing checked and show everything by
default; as soon as at least one box is checked, we show only the
checked categories.